fix: defer product type migration to 6.8 upgrade
