+++ .github/workflows/integration.yml
@@ -599,6 +599,10 @@ jobs:
           DATABASE_URL: "mysql://root@127.0.0.1:3306/shopware_test"
         run: php .github/bin/blue-green-check.php
 
+      - name: Run blue-green integration tests (non-nightly)
+        if: inputs.nightly != 'true'
+        run: php -d memory_limit=-1 vendor/bin/phpunit --filter "GrantDownloadAccessActionTest" --stop-on-error --stop-on-failure
+
   update-66-67:
       name: "PHP update 6.6 -> 6.7"
       runs-on: ubuntu-24.04
